+++ framework/db/ColumnSchemaBuilder.php
@@ -147,13 +147,12 @@ class ColumnSchemaBuilder extends Object
 
     /**
      * Builds the not null constraint for the column.
-     * @return string returns 'NOT NULL' if [[isNotNull]] is true, otherwise it returns an empty string.
+     * @return string returns 'NOT NULL' if [[isNotNull]] is true, 'NULL' otherwise.
      */
     protected function buildNotNullString()
     {
-        return $this->isNotNull ? ' NOT NULL' : '';
+        return $this->isNotNull ? ' NOT NULL' : ' NULL ';
     }
-
 
     /**
      * Builds the unique constraint for the column.